+++ CHANGELOG.md
@@ -15,10 +15,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Lautern real duration now shown in summary
 
 ### Changed
+
 - Fixed summary bugs:
   - Last rast and notes are now displayed in summary
   - Water for sugar solution is now shown in the pre-bottling part
     - Added `bottling_water` column to `summaries` db for persistent storage
+
+### Removed
+- Wrong assumption to always add 1L to the recipe volume to take into consideration yeast loss. This must be done in the recipe not here 
 
 ## [2.1.0] - 2026-02-28